refactor(openmetrics): break up OpenMetricsPlugin god-class into plugin/ and xostor/collectors
